+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/sources/NodeSource.scala
@@ -28,25 +28,43 @@ import uk.co.gresearch.spark.dgraph.connector.encoder.{TypedNodeEncoder, WideNod
 import uk.co.gresearch.spark.dgraph.connector.executor.DgraphExecutorProvider
 import uk.co.gresearch.spark.dgraph.connector.model.NodeTableModel
 import uk.co.gresearch.spark.dgraph.connector.partitioner.PartitionerProvider
+import scala.collection.JavaConverters._
 
 class NodeSource() extends TableProviderBase
   with TargetsConfigParser with SchemaProvider
   with ClusterStateProvider with PartitionerProvider {
 
-  def assertOptions(options: CaseInsensitiveStringMap): Unit = {
+  /**
+   * Sets the number of predicates per partition to max int when predicate partitioner is used
+   * in conjunction with wide node mode. Otherwise wide nodes cannot be properly loaded.
+   *
+   * @param options original options
+   * @return modified options
+   */
+  def adjustOptions(options: CaseInsensitiveStringMap): CaseInsensitiveStringMap = {
     if (getStringOption(NodesModeOption, options).contains(NodesModeWideOption) &&
-      getStringOption(PartitionerOption, options).contains(PredicatePartitionerOption)) {
-      throw new IllegalArgumentException("wide nodes cannot be read with predicate partitioner")
+      getStringOption(PartitionerOption, options).forall(_.startsWith(PredicatePartitionerOption))) {
+      if (getIntOption(PredicatePartitionerPredicatesOption, options).exists(_ != PredicatePartitionerPredicatesDefault)) {
+        println("WARN: predicate partitioner enforced to a single partition to support wide node source")
+      }
+
+      new CaseInsensitiveStringMap(
+        (options.asScala.filterKeys(!_.equalsIgnoreCase(PredicatePartitionerPredicatesOption)) ++
+          Map(PredicatePartitionerPredicatesOption -> Int.MaxValue.toString)
+          ).asJava
+      )
+    } else {
+      options
     }
   }
 
   override def shortName(): String = "dgraph-nodes"
 
   override def inferSchema(options: CaseInsensitiveStringMap): StructType = {
-    assertOptions(options)
-    val targets = getTargets(options)
+    val adjustedOptions = adjustOptions(options)
+    val targets = getTargets(adjustedOptions)
     val schema = getSchema(targets).filter(_.typeName != "uid")
-    getNodeMode(options) match {
+    getNodeMode(adjustedOptions) match {
       case Some(NodesModeTypedOption) => TypedNodeEncoder.schema()
       case Some(NodesModeWideOption) => WideNodeEncoder.schema(schema.predicateMap)
       case Some(mode) => throw new IllegalArgumentException(s"Unknown node mode: ${mode}")
@@ -61,13 +79,13 @@ class NodeSource() extends TableProviderBase
                         partitioning: Array[Transform],
                         properties: util.Map[String, String]): Table = {
     val options = new CaseInsensitiveStringMap(properties)
-    assertOptions(options)
+    val adjustedOptions = adjustOptions(options)
 
-    val targets = getTargets(options)
+    val targets = getTargets(adjustedOptions)
     val schema = getSchema(targets).filter(_.typeName != "uid")
     val clusterState = getClusterState(targets)
-    val partitioner = getPartitioner(schema, clusterState, options)
-    val nodeMode = getNodeMode(options)
+    val partitioner = getPartitioner(schema, clusterState, adjustedOptions)
+    val nodeMode = getNodeMode(adjustedOptions)
     val encoder = nodeMode match {
       case Some(NodesModeTypedOption) => TypedNodeEncoder(schema.predicateMap)
       case Some(NodesModeWideOption) => WideNodeEncoder(schema.predicateMap)
@@ -75,7 +93,7 @@ class NodeSource() extends TableProviderBase
       case None => TypedNodeEncoder(schema.predicateMap)
     }
     val execution = DgraphExecutorProvider()
-    val chunkSize = getIntOption(ChunkSizeOption, options)
+    val chunkSize = getIntOption(ChunkSizeOption, adjustedOptions)
     val model = NodeTableModel(execution, encoder, chunkSize)
     new TripleTable(partitioner, model, clusterState.cid)
   }